+++ nbproject/project.properties
@@ -8,6 +8,8 @@ file.reference.jackson-core-2.8.7.jar=lib/jackson-core-2.8.7.jar
 file.reference.jackson-databind-2.8.7.jar=lib/jackson-databind-2.8.7.jar
 file.reference.jdom-1.0.jar=lib/jdom-1.0.jar
 file.reference.slf4j-api-1.7.24.jar=lib/slf4j-api-1.7.24.jar
+file.reference.taglibs-standard-impl-1.2.5.jar=lib/taglibs-standard-impl-1.2.5.jar
+file.reference.taglibs-standard-spec-1.2.5.jar=lib/taglibs-standard-spec-1.2.5.jar
 j2ee.platform.classpath=${j2ee.server.home}/lib/annotations-api.jar:${j2ee.server.home}/lib/catalina-ant.jar:${j2ee.server.home}/lib/catalina-ha.jar:${j2ee.server.home}/lib/catalina-storeconfig.jar:${j2ee.server.home}/lib/catalina-tribes.jar:${j2ee.server.home}/lib/catalina.jar:${j2ee.server.home}/lib/ecj-4.6.3.jar:${j2ee.server.home}/lib/el-api.jar:${j2ee.server.home}/lib/jasper-el.jar:${j2ee.server.home}/lib/jasper.jar:${j2ee.server.home}/lib/jaspic-api.jar:${j2ee.server.home}/lib/jsp-api.jar:${j2ee.server.home}/lib/servlet-api.jar:${j2ee.server.home}/lib/tomcat-api.jar:${j2ee.server.home}/lib/tomcat-coyote.jar:${j2ee.server.home}/lib/tomcat-dbcp.jar:${j2ee.server.home}/lib/tomcat-i18n-es.jar:${j2ee.server.home}/lib/tomcat-i18n-fr.jar:${j2ee.server.home}/lib/tomcat-i18n-ja.jar:${j2ee.server.home}/lib/tomcat-jdbc.jar:${j2ee.server.home}/lib/tomcat-jni.jar:${j2ee.server.home}/lib/tomcat-util-scan.jar:${j2ee.server.home}/lib/tomcat-util.jar:${j2ee.server.home}/lib/tomcat-websocket.jar:${j2ee.server.home}/lib/websocket-api.jar
 #Mon Jan 29 22:09:39 MST 2018
 javadoc.splitindex=true
@@ -82,7 +84,9 @@ javac.classpath=\
     ${file.reference.jackson-core-2.8.7.jar}:\
     ${file.reference.jackson-databind-2.8.7.jar}:\
     ${file.reference.slf4j-api-1.7.24.jar}:\
-    ${file.reference.jdom-1.0.jar}
+    ${file.reference.jdom-1.0.jar}:\
+    ${file.reference.taglibs-standard-impl-1.2.5.jar}:\
+    ${file.reference.taglibs-standard-spec-1.2.5.jar}
 javadoc.noindex=false
 webinf.dir=web/WEB-INF
 javadoc.reference.API-OMDB-1.5.jar=lib/API-OMDB-1.5-javadoc.jar
